+++ server/src/main/java/pl/edu/agh/to/reaktywni/image/ImageController.java
@@ -1,7 +1,9 @@
 package pl.edu.agh.to.reaktywni.image;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -18,7 +20,8 @@ public class ImageController {
 
     @GetMapping("/{id}")
     public Mono<Image> getImage(@PathVariable int id) {
-        return imageService.getImage(id);
+        return imageService.getImage(id)
+                .switchIfEmpty(Mono.error(new ResponseStatusException(HttpStatus.NOT_FOUND, "Image not found")));
     }
 
     @GetMapping("/thumbnails")
@@ -26,14 +29,10 @@ public class ImageController {
         return imageService.getThumbnails();
     }
 
-    @GetMapping("/count")
-    public long getImagesCount() {
-        return imageService.getImagesCount();
-    }
-
     @GetMapping("/thumbnails/count")
     public Mono<Long> getThumbnailsCount() {
-        return imageService.getThumbnailsCount();
+        return imageService.getThumbnailsCount()
+                .switchIfEmpty(Mono.error(new IllegalStateException("Cannot get thumbnails count")));
     }
 
     @PostMapping(consumes = MediaType.APPLICATION_NDJSON_VALUE, produces = MediaType.APPLICATION_NDJSON_VALUE)