+++ server/src/main/java/pl/edu/agh/to/reaktywni/image/ImageState.java
@@ -2,6 +2,7 @@ package pl.edu.agh.to.reaktywni.image;
 
 
 public enum ImageState {
+    PENDING,
     SUCCESS,
     FAILURE
 }